Replaced template orbits by orbital parameters factory.

Closes #1867 and #1974

See merge request orekit/orekit!1253

+++ src/test/java/org/orekit/files/ccsds/ndm/NdmTestUtils.java
@@ -346,7 +346,7 @@ public class NdmTestUtils {
     }
 
     public static void checkDouble(final Double original, final Double rebuilt) {
-        Assertions.assertTrue(Precision.equalsIncludingNaN(original.doubleValue(), rebuilt.doubleValue(), ULPS));
+        Assertions.assertTrue(Precision.equalsIncludingNaN(original, rebuilt, ULPS));
     }
